list_gen_v2/requirements.txt
# Core dependencies for SQL generation and execution
pandas>=1.5.0
pyodbc>=4.0.0
openpyxl>=3.1.0
PyYAML>=6.0
Flask>=2.0.0
openai>=1.0.0

# Dependencies likely already present from existing system
# (including these for completeness)
requests>=2.28.0
